Fixed Scanner escape sequence, Added Scanner tests

+++ IFJ-FreeBASIC2-compiler/string.c
@@ -19,7 +19,7 @@ void AddToString(char c, tToken *Token, int LenghtOfString)
 
 void ReallocString(tToken *Token, int LenghtOfString)
 {
-	if ((Token->String = mrealloc(Token->String, sizeof(char) * LenghtOfString)) == NULL)
+	if ((Token->String = mrealloc(Token->String, sizeof(char) * (size_t) LenghtOfString)) == NULL)
 	{
 		mfreeall();
 		ERR_CODE code = INTERNAL_ERR;
